+++ backend/src/main/java/reviewme/reviewgroup/domain/ReviewGroup.java
@@ -10,7 +10,7 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import reviewme.reviewgroup.domain.exception.InvalidProjectNameLengthException;
-import reviewme.reviewgroup.domain.exception.InvalidRevieweeNameLengthException;
+import reviewme.reviewgroup.service.exception.InvalidRevieweeNameLengthException;
 
 @Entity
 @Table(name = "review_group")
