+++ backend/src/main/java/reviewme/reviewgroup/service/exception/ReviewGroupNotFoundByReviewException.java
@@ -1,4 +1,4 @@
-package reviewme.reviewgroup.domain.exception;
+package reviewme.reviewgroup.service.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import reviewme.global.exception.DataInconsistencyException;
